+++ app/app.js
@@ -12,6 +12,7 @@ app.disable('x-powered-by');
 app.use(helmet.frameguard({ action: 'deny' }));
 app.use(helmet.noSniff());
 app.use(helmet.ieNoOpen());
+app.use(helmet.xssFilter());
 app.use(helmet.referrerPolicy({ policy: 'no-referrer' }));
 app.use(helmet.hsts({
       maxAge: 31536000000,